+++ CHANGES.md
@@ -5,6 +5,9 @@ Unreleased
   0.7. `(mode native)` has been deprecated in favour of detection from the
   configuration of Coq. (#6409, @Alizter)
 
+- Add support for `coqffi` with the `coqffi` stanza (#6248, fixes #3995,
+  @Alizter)
+
 3.6.0 (2022-11-14)
 ------------------
 
@@ -68,7 +71,6 @@ Unreleased
 
 3.5.0 (2022-10-19)
 ------------------
-
 - macOS: Handle unknown fsevents without crashing (#6217, @rgrinberg)
 
 - Enable file watching on MacOS SDK < 10.13. (#6218, @rgrinberg)